nim-excercise/requirements.txt
nemoguardrails
load-dotenv